optimize by downloading once

+++ survey_all.rb
@@ -1,4 +1,7 @@
+require 'fileutils'
+
 `curl https://raw.githubusercontent.com/tenderlove/gem_survey/master/survey.rb | ruby`
+`curl https://raw.githubusercontent.com/tenderlove/gem_survey/master/survey.rb > /tmp/survey.rb`
 
 ruby_project_paths = Dir['/**/Gemfile'].map { |s| s.split('/')[0..-2].join('/') }
 
@@ -6,9 +9,15 @@ ruby_project_paths.each { |path|
   puts "Running in #{path}"
 
   Dir.chdir(path)
-  `curl https://raw.githubusercontent.com/tenderlove/gem_survey/master/survey.rb | bundle exec ruby`
-  puts "Cleaning"
-  File.delete('survey.rb')
+
+  begin
+    FileUtils.cp('/tmp/survey.rb', "#{path}/survey.rb")
+    `bundle exec ruby ./survey.rb 2>/dev/null`
+  rescue => e
+    puts e
+  ensure
+    File.delete('survey.rb') if File.exist?('survey.rb')
+  end
 }
 
 puts "Thanks for taking the time to give back to Ruby!"
